+++ CHANGELOG.md
@@ -8,6 +8,10 @@ All notable changes to this project will be documented in this file. The format 
 - Add `Mesh.clear_points_without_cells()` to clear the list of points without cells (useful for center-points of multi-point constraints).
 - Release FElupe on conda-forge, starting with v9.2.0.
 
+### Changed
+- Change the required setuptools-version in the build-system table of `pyproject.toml` to match PEP639 (setuptools>=77.0.3).
+- Change the labels to well-known labels for the URLs in `pyproject.toml`.
+
 ### Fixed
 - Fix the declaration of the (spdx identifier) license and license-file in `pyproject.toml`.